+++ tests/cpcm/cpcm_gepol-point_from-file.cpp
@@ -57,9 +57,9 @@ TEST_CASE("Test solver for the C-PCM for a point charge and a restarted GePol ca
     solver.buildSystemMatrix(cavity, gf_i, gf_o);
 
     double charge = 8.0;
-    size_t size = cavity.size();
+    int size = cavity.size();
     Eigen::VectorXd fake_mep = Eigen::VectorXd::Zero(size);
-    for (size_t i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         Eigen::Vector3d center = cavity.elementCenter(i);
         double distance = center.norm();
         fake_mep(i) = charge / distance;
