+++ tests/gepol/gepol_C2H4_D2h.cpp
@@ -56,7 +56,7 @@ SCENARIO("GePol cavity for the C2H4 molecule in D2h symmetry", "[gepol][gepol_C2
             THEN("the size of the cavity is")
             {
                 int size = 576;
-                size_t actualSize = cavity.size();
+                int actualSize = cavity.size();
                 REQUIRE(size == actualSize);
             }
             /*! \class GePolCavity
@@ -86,7 +86,7 @@ SCENARIO("GePol cavity for the C2H4 molecule in D2h symmetry", "[gepol][gepol_C2
                 Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
                 Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
                 double actualVolume = 0;
-                for ( size_t i = 0; i < cavity.size(); ++i ) {
+                for ( int i = 0; i < cavity.size(); ++i ) {
                     actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                                 i));
                 }
@@ -109,7 +109,7 @@ SCENARIO("GePol cavity for the C2H4 molecule in D2h symmetry", "[gepol][gepol_C2
             THEN("the size of the cavity is")
             {
                 int size = 576;
-                size_t actualSize = cavity.size();
+                int actualSize = cavity.size();
                 REQUIRE(size == actualSize);
             }
             /*! \class GePolCavity
@@ -139,7 +139,7 @@ SCENARIO("GePol cavity for the C2H4 molecule in D2h symmetry", "[gepol][gepol_C2
                 Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
                 Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
                 double actualVolume = 0;
-                for ( size_t i = 0; i < cavity.size(); ++i ) {
+                for ( int i = 0; i < cavity.size(); ++i ) {
                     actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                                 i));
                 }
