+++ tests/gepol/gepol_C6H6.cpp
@@ -51,7 +51,7 @@ TEST_CASE("GePol cavity for the benzene molecule", "[gepol][gepol_C6H6]")
     SECTION("Test size")
     {
         int size = 644;
-        size_t actualSize = cavity.size();
+        int actualSize = cavity.size();
         REQUIRE(size == actualSize);
     }
 
@@ -74,7 +74,7 @@ TEST_CASE("GePol cavity for the benzene molecule", "[gepol][gepol_C6H6]")
         Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
         Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
         double actualVolume = 0;
-        for ( size_t i = 0; i < cavity.size(); ++i ) {
+        for ( int i = 0; i < cavity.size(); ++i ) {
             actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                         i));
         }
