+++ tests/gepol/gepol_CH3+_Cs.cpp
@@ -50,7 +50,7 @@ TEST_CASE("GePol cavity for the CH3+ molecule in Cs symmetry", "[gepol][gepol_CH
     SECTION("Test size")
     {
         int size = 384;
-        size_t actualSize = cavity.size();
+        int actualSize = cavity.size();
         REQUIRE(size == actualSize);
     }
 
@@ -83,7 +83,7 @@ TEST_CASE("GePol cavity for the CH3+ molecule in Cs symmetry", "[gepol][gepol_CH
         Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
         Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
         double actualVolume = 0;
-        for ( size_t i = 0; i < cavity.size(); ++i ) {
+        for ( int i = 0; i < cavity.size(); ++i ) {
             actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                         i));
         }
