+++ tests/gepol/gepol_H3+_C2v.cpp
@@ -55,7 +55,7 @@ SCENARIO("GePol cavity for the H3+ molecule in C2v symmetry", "[gepol][gepol_H3+
             THEN("the size of the cavity is")
             {
                 int size = 312;
-                size_t actualSize = cavity.size();
+                int actualSize = cavity.size();
                 REQUIRE(size == actualSize);
             }
             /*! \class GePolCavity
@@ -85,7 +85,7 @@ SCENARIO("GePol cavity for the H3+ molecule in C2v symmetry", "[gepol][gepol_H3+
                 Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
                 Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
                 double actualVolume = 0;
-                for ( size_t i = 0; i < cavity.size(); ++i ) {
+                for ( int i = 0; i < cavity.size(); ++i ) {
                     actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                                 i));
                 }
@@ -108,7 +108,7 @@ SCENARIO("GePol cavity for the H3+ molecule in C2v symmetry", "[gepol][gepol_H3+
             THEN("the size of the cavity is")
             {
                 int size = 288;
-                size_t actualSize = cavity.size();
+                int actualSize = cavity.size();
                 REQUIRE(size == actualSize);
             }
             /*! \class GePolCavity
@@ -138,7 +138,7 @@ SCENARIO("GePol cavity for the H3+ molecule in C2v symmetry", "[gepol][gepol_H3+
                 Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
                 Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
                 double actualVolume = 0;
-                for ( size_t i = 0; i < cavity.size(); ++i ) {
+                for ( int i = 0; i < cavity.size(); ++i ) {
                     actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                                 i));
                 }
