+++ tests/gepol/gepol_NH3.cpp
@@ -52,7 +52,7 @@ TEST_CASE("GePol cavity for an ammonia molecule", "[gepol][gepol_NH3]")
     SECTION("Test size")
     {
         int size = 230;
-        size_t actualSize = cavity.size();
+        int actualSize = cavity.size();
         REQUIRE(size == actualSize);
     }
 
@@ -75,7 +75,7 @@ TEST_CASE("GePol cavity for an ammonia molecule", "[gepol][gepol_NH3]")
         Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
         Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
         double actualVolume = 0;
-        for ( size_t i = 0; i < cavity.size(); ++i ) {
+        for ( int i = 0; i < cavity.size(); ++i ) {
             actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                         i));
         }
