+++ tests/gepol/gepol_point.cpp
@@ -53,7 +53,7 @@ SCENARIO("GePol cavity for a single sphere", "[gepol][gepol_point]")
       THEN("the size of the cavity is")
       {
         int size = 32;
-        size_t actualSize = cavity.size();
+        int actualSize = cavity.size();
         REQUIRE(size == actualSize);
       }
       /*! \class GePolCavity
@@ -74,7 +74,7 @@ SCENARIO("GePol cavity for a single sphere", "[gepol][gepol_point]")
         Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
         Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
         double actualVolume = 0;
-        for ( size_t i = 0; i < cavity.size(); ++i ) {
+        for ( int i = 0; i < cavity.size(); ++i ) {
           actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                 i));
         }
@@ -100,7 +100,7 @@ SCENARIO("GePol cavity for a single sphere", "[gepol][gepol_point]")
       THEN("the size of the cavity is")
       {
         int size = 32;
-        size_t actualSize = cavity.size();
+        int actualSize = cavity.size();
         REQUIRE(size == actualSize);
       }
       /*! \class GePolCavity
@@ -121,7 +121,7 @@ SCENARIO("GePol cavity for a single sphere", "[gepol][gepol_point]")
         Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
         Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
         double actualVolume = 0;
-        for ( size_t i = 0; i < cavity.size(); ++i ) {
+        for ( int i = 0; i < cavity.size(); ++i ) {
           actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                 i));
         }
